core/standards/XMLHttpRequest/send-entity-body-none-explicit.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send(null) - no entity body, explicit content-type</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        var client = new XMLHttpRequest()
        client.open("POST", "resources/content.php", false);
	var content_type = 'application/x-foo';
	client.setRequestHeader('Content-Type', content_type);
        client.send(null);
        assert_equals(client.getResponseHeader("x-request-content-length"), "0");
        assert_equals(client.getResponseHeader("x-request-content-type"), content_type);
      })
    </script>
  </body>
</html>
